Changes: - changed class that deals with ES from search manager to snippet_snapshot_manager - extracted logic from search_view to put it in SearchManager class
